refactor: decompose backend analyzer and extract move cell widgets
